chore: :technologist: add eslint, prettier, jsdoc and typescript for linting and quality control
This is just an initial setup of this tooling. We'll be adding more
rules and configuration as we go. Ideally, all configuration should be
in `.js` where possible and `.json` as a last resort. With configuration
files in `.js`, we can add JSDoc comments to explain our choices and
use type annotations to get auto-completion and type checking, since
most build tooling has some kind of documentation for the configuration
they support.